Fixed Wide Screen Viewing
The previous usage of container seemed to cause quirky behavior with
really wide screens, due to the difference in stretching between the
header and the list items.

Adding container to the body seems to have prompted a fixed width
layout. Not sure why honestly, Neat is still a bit of a mystery to me,
but hey.. my HTML is clean, so yay lol.